% Generated by build script
\name{neon_product}
\alias{neon_product}
\title{Fetch details for a specific NEON data product}
\usage{
neon_product()
}
\arguments{
\item{code}{NEON product code (e.g. "DP1.10003.001")}
}
\value{tibble: one row with productCode, productName, productDescription,}
\description{Fetch details for a specific NEON data product}
