% Generated by build script
\name{neon_site}
\alias{neon_site}
\title{Get detailed info for a single NEON site}
\usage{
neon_site()
}
\arguments{
\item{site_code}{NEON site code (e.g. "HARV", "BART")}
}
\value{tibble: siteCode, siteDescription, siteType, stateName,}
\description{Get detailed info for a single NEON site}
